requirements.txt
argparse
trimesh
pyopengl
PyQt5
pandas
openpyxl
torch>=0.4.1
torchvision>=0.2.1
numpy>=1.15.4
scipy>=1.1.0
matplotlib==3.0.2
dlib==19.5.0 
opencv-python>=3.4.3.18
